refactor: colocate creds tests and fix names
